+++ blender-qmm/mats/TexturizerGroupNode.py
@@ -16,8 +16,15 @@ class TexturizerGroup(bpy.types.Operator):
             group_in = texturizer_group.nodes.new('NodeGroupInput')
             group_in.location = (-800, 0)
             texturizer_group.inputs.new('NodeSocketColor', 'Color')
+            texturizer_group.inputs.new('NodeSocketFloat', 'Saturation')
+            texturizer_group.inputs.new('NodeSocketFloat', 'Mix')
+            texturizer_group.inputs.new('NodeSocketFloat', 'Bump')
             texturizer_group.inputs.new('NodeSocketVector', 'Vector')
-            texturizer_group.inputs[1].hide_value = True
+            texturizer_group.inputs[4].hide_value = True
+
+            texturizer_group.inputs[1].default_value = 0.5
+            texturizer_group.inputs[2].default_value = 0.5
+            texturizer_group.inputs[3].default_value = 0.1
 
             #groupoutput
             group_out = texturizer_group.nodes.new('NodeGroupOutput')
@@ -50,14 +57,17 @@ class TexturizerGroup(bpy.types.Operator):
             #noisetexture
             n_tex = texturizer_group.nodes.new('ShaderNodeTexNoise')
             n_tex.location = (-600,-100)
-            n_tex.inputs[2].default_value = 512
-            n_tex.inputs[3].default_value = 0
-            n_tex.inputs[4].default_value = 0
+            n_tex.inputs[2].default_value = 256
+            n_tex.inputs[3].default_value = 2.0
+            n_tex.inputs[4].default_value = 0.5
 
             links = texturizer_group.links.new
 
             links(group_in.outputs[0], n_mix_rgb.inputs[1])
-            links(group_in.outputs[1], n_tex.inputs[0])
+            links(group_in.outputs[1], n_hsl.inputs[1])
+            links(group_in.outputs[2], n_mix_rgb.inputs[0])
+            links(group_in.outputs[3], n_bump.inputs[0])
+            links(group_in.outputs[4], n_tex.inputs[0])
             links(n_tex.outputs[0], n_mr.inputs[0])
             links(n_tex.outputs[1], n_hsl.inputs[4])
             links(n_hsl.outputs[0], n_mix_rgb.inputs[2])