+++ blender-qmm/mats/noble/Gold.py
@@ -9,7 +9,7 @@ def ShowMessageBox(message = "", title = "", icon = 'INFO'):
 
 #GoldShaderOperator
 class QMMGold(bpy.types.Operator):
-    """Add/Apply Gold Metallic Material to Selected Object (or Scene)"""
+    """Add/Apply Gold Material to Selected Object (or Scene)"""
     bl_label = "QMM Gold Shader"
     bl_idname = 'shader.qmm_gold_m_operator'
     def execute(self, context):